requirements.txt
fastapi
uvicorn
websockets
python-dotenv
httpx
numpy
sentence-transformers